render: Move `common_impl.h` into `clx_render.cpp`
We no longer have other renderers.
